+++ raw_tx.h
@@ -6,6 +6,6 @@
 /* Transmit the packet in the packetbuf, using
  * the routing subsystem to calculate the next hop address.
  */
-int SN_Forward_Packetbuf(uint16_t source, uint16_t destination);
+int8_t SN_Forward_Packetbuf(uint16_t source, uint16_t destination);
 
 #endif //STARFISHNET_NONQUEUED_TRANSMISSION_H
